relocate some input fields to action type for easy understanding of input type
